third_party/WebKit/LayoutTests/fast/gradients/css3-gradient-parsing-expected.txt

linear-gradient

PASS testGradient("background-image: linear-gradient(black 0%", "background-image") is "none"
PASS testGradient("background-image: linear-gradient(top)", "background-image") is "none"
PASS testGradient("background-image: linear-gradient(10deg)", "background-image") is "none"
PASS testGradient("background-image: linear-gradient(top center, black 0%)", "background-image") is "none"
PASS testGradient("background-image: linear-gradient(10px 20%, black 0%)", "background-image") is "none"
PASS testGradient("background-image: linear-gradient(left left)", "background-image") is "none"
PASS testGradient("background-image: linear-gradient(left center)", "background-image") is "none"
PASS testGradient("background-image: linear-gradient(left 10deg, black 0%)", "background-image") is "none"
PASS testGradient("background-image: linear-gradient(black)", "background-image") is "none"
PASS testGradient("background-image: linear-gradient(right bottom, black, white)", "background-image") is "none"
PASS testGradient("background-image: linear-gradient(black, white)", "background-image") is "linear-gradient(black, white)"
PASS testGradient("background-image: linear-gradient(black 0, white)", "background-image") is "linear-gradient(black 0px, white)"
PASS testGradient("background-image: linear-gradient(black 0%, white)", "background-image") is "linear-gradient(black 0%, white)"
PASS testGradient("background-image: linear-gradient(black 0%, white)", "background-image") is "linear-gradient(black 0%, white)"
PASS testGradient("background-image: linear-gradient(black, white)", "background-image") is "linear-gradient(black, white)"
PASS testGradient("background-image: linear-gradient(black 10px, white 20px)", "background-image") is "linear-gradient(black 10px, white 20px)"
PASS testGradient("background-image: linear-gradient(to bottom right, black 0%, white)", "background-image") is "linear-gradient(to right bottom, black 0%, white)"
PASS testGradient("background-image: linear-gradient(black 0%, white)", "background-image") is "linear-gradient(black 0%, white)"
PASS testGradient("background-image: linear-gradient(10deg, black 0%, white)", "background-image") is "linear-gradient(10deg, black 0%, white)"
repeating-linear-gradient

PASS testGradient("background-image: repeating-linear-gradient(black, white)", "background-image") is "repeating-linear-gradient(black, white)"
radial-gradient

PASS testGradient("background-image: radial-gradient(white, black)", "background-image") is "radial-gradient(white, black)"
PASS testGradient("background-image: radial-gradient(at bottom right, white, black)", "background-image") is "radial-gradient(at right bottom, white, black)"
PASS testGradient("background-image: radial-gradient(ellipse farthest-corner, white, black)", "background-image") is "radial-gradient(white, black)"
PASS testGradient("background-image: radial-gradient(circle closest-corner, white, black)", "background-image") is "radial-gradient(circle closest-corner, white, black)"
PASS testGradient("background-image: radial-gradient(circle, white, black)", "background-image") is "radial-gradient(circle, white, black)"
PASS testGradient("background-image: radial-gradient(circle closest-side, white, black)", "background-image") is "radial-gradient(circle closest-side, white, black)"
PASS testGradient("background-image: radial-gradient(circle closest-side at top, white, black)", "background-image") is "radial-gradient(circle closest-side at center top, white, black)"
PASS testGradient("background-image: radial-gradient(circle closest-side at top left, white, black)", "background-image") is "radial-gradient(circle closest-side at left top, white, black)"
PASS testGradient("background-image: radial-gradient(circle closest-side at 10px 20%, white, black)", "background-image") is "radial-gradient(circle closest-side at 10px 20%, white, black)"
PASS testGradient("background-image: radial-gradient(10px, 20%, circle closest-side, white, black)", "background-image") is "none"
PASS testGradient("background-image: radial-gradient(circle 10px 20%, circle closest-side, white, black)", "background-image") is "none"
PASS testGradient("background-image: radial-gradient(circle 10px, circle closest-side, white, black)", "background-image") is "none"
PASS testGradient("background-image: radial-gradient(10px, white, black)", "background-image") is "radial-gradient(10px, white, black)"
PASS testGradient("background-image: radial-gradient(10px 10px, white, black)", "background-image") is "radial-gradient(10px 10px, white, black)"
PASS testGradient("background-image: radial-gradient(ellipse closest-corner, white, black)", "background-image") is "radial-gradient(closest-corner, white, black)"
PASS testGradient("background-image: radial-gradient(circle closest-side, white, black)", "background-image") is "radial-gradient(circle closest-side, white, black)"
repeating-radial-gradient

PASS testGradient("background-image: repeating-radial-gradient(white, black)", "background-image") is "repeating-radial-gradient(white, black)"
PASS successfullyParsed is true

TEST COMPLETE
